index.html: Revamps back into single file
I decided to leverage Gemini a bit more,
which was shoving everything into one
massive file. I ended up being ok with it
because it was making a lot of progress
in terms of building out the functionality.

I figured I can just minify it before
hosting on GH pages and if/when I decide to
go for a full version simulation then I can
consider pulling things out into more logical
files.

That being said, Gemini ended up hallucinating
real bad and also lost access to a good state
so I'm going to be moving to cursor to see how
it fares. This is effectively the last working
state I had copied over from Gemini before it
completely shit itself. I also updated the
README to capture some of the features made,
change in priorities in terms of feature
building (drills/exercises rather than full sim)
and some of the known bugs that I'll have to
re-address and iron out.