index.html: Tweaks to copy, visuals
Adds small copy tweaks (footer) and
attempts to fix the visuals on
Immobilism.

Also updates the README to better
reflect how things are. I think
we're pretty close to a v1! Mostly
just want to include a few 3 unit
variations before I think it's
ready to go public. Depending on
reception I might make the full
sim.

Shout out to Gemini and Cursor
for getting this far so fast