README.md: Cleans up readme
There were lots of MVP tasks for the initial set
up of drills that are complete. I feel pretty
ok with cleaning up old tasks that have been
done and make it a bit clearer what's coming
next as well as which known bugs/edge cases are
being looked at.