index.html: Bug fixes + small refactor
This commit pulls the grid state into an object to act as
an enum (with more human readable states). In terms of
bug fixes, we added the 1AP cost to immobilism (validated
from video) and prevent the use of Double Leap behind an
enemy (LoS, although this would probably result in a loss
anyway so no real reason to keep it around).

Outside of the scope of this commit - there's one more
major bug that involves enemies being merged onto the same
square, but before this can be addressed I need to learn
what the game actually does in these particular scenarios.